refactor: common -> shared
