+++ src/shared/game.js
