+++ src/shared/util/vector.js
